refactor: 컴포넌트명을 수정합니다.
