+++ src/funnels/on-boarding/Step2.tsx
